+++ Processing/HandTieArduinoSystemOnProcessing/SerialNotifier.pde
@@ -2,13 +2,13 @@ public interface SerialNotifier{
    public void registerForSerialListener(SerialListener listener);
    public void removeSerialListener(SerialListener listener);
    
-   public void notifyAllWithAnalogVals(int [] values);
-   public void notifyAllWithCaliVals(int [] values);
-   public void notifyAllWithTargetAnalogValsMinAmp(int [] values);
-   public void notifyAllWithTargetAnalogValsWithAmp(int [] values);
-   public void notifyAllWithBridgePotPosVals(int [] values);
-   public void notifyAllWithAmpPotPosVals(int [] values);
-   public void notifyAllWithCalibratingValsMinAmp(int [] values);
-   public void notifyAllWithCalibratingValsWithAmp(int [] values);
+   public void notifyAllWithAnalogVals(float [] values);
+   public void notifyAllWithCaliVals(float [] values);
+   public void notifyAllWithTargetAnalogValsMinAmp(float [] values);
+   public void notifyAllWithTargetAnalogValsWithAmp(float [] values);
+   public void notifyAllWithBridgePotPosVals(float [] values);
+   public void notifyAllWithAmpPotPosVals(float [] values);
+   public void notifyAllWithCalibratingValsMinAmp(float [] values);
+   public void notifyAllWithCalibratingValsWithAmp(float [] values);
    public void notifyAllWithReceiveRecordSignal();
 }
